+++ NEWS.md
@@ -1,4 +1,5 @@
 # spant 3.4.0
+* Added a function to read DKD's sLASER MoCo log file.
 
 # spant 3.3.0
 * Added voxel render on MRI images to fit_svs.